Title: 길이 단위 환산하기, Time: 145ms, Memory: 11MB, Status: Passed - Codetree

+++ 260128/길이 단위 환산하기/convert-length-units.java
@@ -5,7 +5,7 @@ public class Main {
         // Please write your code here.
         Scanner sc=new Scanner(System.in);
 
-        double N=sc.nextInt();
+        double N=sc.nextFloat();
         double ft=30.48;
 
         System.out.printf("%.1f",N*ft);
